<<<<<<< HEAD
\name{optSelect-package}
\alias{optSelect-package}
\alias{optSelect}
\docType{package}
\title{optSelect}
\description{
optSelect, a multi agent-based stochastic optimization approach for ensemble feature selection. Stage one involves function perturbation, where ranked list of features are generated using different feature selection methods and stage two involves data perturbation, where feature selection is performed within randomly selected learning sets of the training data. The agents are assigned to different behavior states and move according to a binary PSO algorithm. A multi-objective fitness function is used to evaluate the classifica-tion accuracy of the agents. 
}
\details{
\tabular{ll}{
Package: \tab optSelect\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2020-11-01\cr
License: \tab gpl2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Karan Uppal
Maintainer: <kuppal3gt@gmail.com>
}
\keyword{ optSelect }
=======
\name{optSelect_v0.0.1-package}
\alias{optSelect_v0.0.1-package}
\alias{optSelect_v0.0.1}
\docType{package}
\title{
\packageTitle{optSelect_v0.0.1}
}
\description{
\packageDescription{optSelect_v0.0.1}
}
\details{
optSelect is an R package for ensemble feature selection and stability assessment. It uses a multi agent-based stochastic optimization approach for ensemble feature selection. Stage one involves function perturbation, where ranked list of features are generated using different feature selection methods and stage two involves data perturbation, where feature selection is performed within randomly selected learning sets of the training data. The agents are assigned to different behavior states and move according to a binary PSO algorithm. A multi-objective fitness function is used to evaluate the classifica-tion accuracy of the agents. 
}
\author{
\packageAuthor{Karan Uppal}

Maintainer: \packageMaintainer{Karan Uppal}
}
>>>>>>> 53e39c9aa29f120f9c9571d6158c60dfad65abf5
